requirements.txt
Django==6.0.3
Faker==40.11.1
boto3==1.42.76
botocore==1.42.76
crispy-bootstrap5==2026.3
dj-database-url==3.1.2
django-allauth==65.15.0
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-countries==8.2.0
django-crispy-forms==2.6
django-ninja==1.6.2
django-storages==1.14.6
django-webtest==1.9.14
djangorestframework==3.17.1
gunicorn==25.2.0
pillow==12.1.1
psycopg2-binary==2.9.11
pytest-cov==7.1.0
pytest-django==4.12.0
pytest==9.0.2
python-dotenv==1.2.2
stripe==14.4.1
